#fix upload controller logic shidt to controller layer
