+++ src/controllers/user.controller.ts
@@ -1,0 +1,31 @@
+import { Request, Response } from 'express';
+import { StatusCodes } from 'http-status-codes';
+
+async function uploadResumeHandler(req: Request, res: Response) {
+    if (!req.file) {
+        res.status(StatusCodes.BAD_REQUEST).json({
+            success: false,
+            message: 'No file uploaded',
+            data: {},
+            error: {
+                message: 'Bad Request'
+            }
+        });
+    }
+
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const fileUrl = (req.file as any).location;
+
+    res.status(StatusCodes.OK).json({
+        success: true,
+        message: 'Succesfully uploaded the resume',
+        data: {
+            fileUrl
+        },
+        error: {}
+    });
+}
+
+export {
+    uploadResumeHandler
+};
